backend/requirements/base.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# CORS
django-cors-headers>=4.3,<5.0

# Filtering
django-filter>=23.5,<24.0

# Environment variables
python-dotenv>=1.0,<2.0

# Rate limiting
django-ratelimit>=4.1,<5.0
